+++ lib/influxdb-rails.rb
@@ -38,7 +38,11 @@ module InfluxDB
           :port => configuration.influxdb_port,
           :async => configuration.async,
           :use_ssl => configuration.use_ssl,
-          :retry => configuration.retry
+          :retry => configuration.retry,
+          :open_timeout => configuration.open_timeout,
+          :read_timeout => configuration.read_timeout,
+          :max_delay => configuration.max_delay,
+          :time_precision => configuration.time_precision
       end
 
       def configuration
@@ -73,7 +77,7 @@ module InfluxDB
       alias_method :transmit, :report_exception
 
       def handle_action_controller_metrics(name, start, finish, id, payload)
-        timestamp = finish.utc.to_i
+        timestamp = convert_timestamp(finish.utc)
         controller_runtime = ((finish - start)*1000).ceil
         view_runtime = (payload[:view_runtime] || 0).ceil
         db_runtime = (payload[:db_runtime] || 0).ceil
@@ -90,6 +94,7 @@ module InfluxDB
               server: hostname,
               application: configuration.application_name
             },
+            timestamp: timestamp,
           }
 
           client.write_point configuration.series_name_for_view_runtimes, {
@@ -101,6 +106,7 @@ module InfluxDB
               server: hostname,
               application: configuration.application_name
             },
+            timestamp: timestamp,
           }
 
           client.write_point configuration.series_name_for_db_runtimes, {
@@ -112,14 +118,34 @@ module InfluxDB
               server: hostname,
               application: configuration.application_name
             },
+            timestamp: timestamp,
           }
         rescue => e
           log :error, "[InfluxDB::Rails] Unable to write points: #{e.message}"
         end
       end
 
+      def convert_timestamp(ts)
+        case configuration.time_precision
+        when 'ns', nil
+          (ts.to_r * 1e9).to_i
+        when 'u'
+          (ts.to_r * 1e6).to_i
+        when 'ms'
+          (ts.to_r * 1e3).to_i
+        when 's'
+          ts.to_i
+        when 'm'
+          ts.to_i / 60
+        when 'h'
+          ts.to_i / 60 / 60
+        else
+          raise "Invalid time precision: #{configuration.time_precision}"
+        end
+      end
+
       def current_timestamp
-        Time.now.utc.to_i
+        convert_timestamp(Time.now.utc)
       end
 
       def ignorable_exception?(e)
@@ -143,6 +169,15 @@ module InfluxDB
       rescue StandardError => e
         transmit_unless_ignorable(e)
         raise(e)
+      end
+
+      def safely_prepend(module_name, opts = {})
+        return if opts[:to].nil? || opts[:from].nil?
+        if opts[:to].respond_to?(:prepend, true)
+          opts[:to].send(:prepend, opts[:from].const_get(module_name))
+        else
+          opts[:to].send(:include, opts[:from].const_get("Old" + module_name))
+        end
       end
     end
   end